+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/feemarket/FeeMarketException.java
@@ -25,6 +25,11 @@ public class FeeMarketException extends Exception {
         "Invalid block header: basefee should not be present before fee market fork block");
   }
 
+  public static FeeMarketException MissingBaseFeeMarket() {
+    return new FeeMarketException(
+        "Incorrectly configured ProtocolSchedule: BaseFeeBlockBodyValidator requires BaseFeeMarket");
+  }
+
   private FeeMarketException(final String reason) {
     super(reason);
   }
